feat: copy poem to clipboard
