+++ package.json
@@ -26,6 +26,7 @@
     "morgan": "^1.10.0",
     "pg": "^8.7.3",
     "react": "^18.2.0",
+    "react-copy-to-clipboard": "^5.1.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
@@ -66,6 +67,7 @@
   },
   "devDependencies": {
     "@types/lodash": "^4.14.182",
+    "@types/react-copy-to-clipboard": "^5.0.4",
     "eslint-plugin-react": "^7.30.1"
   }
 }
