+++ src/components/Poem/styles.ts
@@ -1,0 +1,10 @@
+import * as React from "react";
+
+export const poemFont: React.CSSProperties = {
+    lineHeight: "1.5em",
+    whiteSpace: "pre-line",
+};
+
+export const poemTitle: React.CSSProperties = {
+    margin: "auto",
+};
