+++ src/components/Poems/styles.ts
@@ -4,12 +4,3 @@ export const poemsBody: React.CSSProperties = {
     marginBottom: "2em",
     marginTop: "auto",
 };
-
-export const poemFont: React.CSSProperties = {
-    lineHeight: "1.5em",
-    whiteSpace: "pre-line",
-};
-
-export const poemTitle: React.CSSProperties = {
-    margin: "auto",
-};
